rename deprecated-signatures to v3-deprecated-signatures, cover json and jsonp cases, unify logic for 2 arguments with status and body
