--- transforms/__test__/deprecated-signatures.spec.ts
@@ -1,3 +1,0 @@
-import { testSpecBuilder } from './util'
-
-testSpecBuilder('deprecated-signatures')
