+++ transforms/__test__/v3-deprecated-signatures.spec.ts
@@ -1,0 +1,3 @@
+import { testSpecBuilder } from './util'
+
+testSpecBuilder('v3-deprecated-signatures')
